# university-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University API",
        "description": "University Management System API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/enrollments": {
            "post": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Enroll a student in a course",
                "description": "ENROLL STUDENT",
                "operationId": "6a2e292dcc494335970a261a2772dd66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id",
                                    "student_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "student_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enrolled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "enrollment_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "course_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "student_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "enrolled_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Already Enrolled or Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/students/{student_id}/courses": {
            "get": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "List courses for a specific student",
                "description": "STUDENT COURSES",
                "operationId": "c9fa0be41ee53f334225b5f4682ccd8c",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "description": "ID of the student",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of enrollments with course details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "enrollment_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "course": {
                                                "properties": {
                                                    "course_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Introduction to Computer Science"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "enrolled_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/courses/{course_id}/students": {
            "get": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Get list of students enrolled in a course (Admin/Instructor)",
                "description": "LIST STUDENTS IN COURSE",
                "operationId": "e5dba241a9d5079c029bb07d507857ad",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of enrolled students",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "enrollment_id": {
                                                "type": "integer"
                                            },
                                            "student": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/enrollments/{enrollment_id}": {
            "delete": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Unenroll a student from a course (Admin only)",
                "description": "UNENROLL STUDENT",
                "operationId": "1b2e3dc5a04589050df0ca5f9c597353",
                "parameters": [
                    {
                        "name": "enrollment_id",
                        "in": "path",
                        "description": "ID of the enrollment to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unenrolled successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Enrollment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}